requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
pydantic>=1.8.2
loguru>=0.5.3
PyYAML>=6.0
openai>=0.27.0
tiktoken>=0.3.0
APScheduler>=3.9.1 